add renderChild prop to TreeNode so we can separate node and container into separate files without circular deps; update effected imports
