+++ frontend/src/components/DiffAnalyzer.tsx
@@ -2,7 +2,7 @@ import CodeEditor from "./CodeEditor";
 import FilterDropdown from "./filterDropdown";
 import { parseDiff } from "../utils/parsingMessageHandlers";
 import { ASTNode, VSCodeAPI } from "../types/typesAndInterfaces";
-import TreeNodeContainer from "./TreeNode";
+import TreeNodeContainer from "./TreeNodeContainer";
 
 interface DiffAnalyzerProps {
   codeSnippet1: string;
