+++ frontend/src/components/TreeNode.tsx
@@ -1,14 +1,12 @@
 import React from "react";
 import { TypeAnnotation } from "./TypeAnnotation";
-import { shouldAutoCollapse } from "../utils/nodeEmphasisHelpers";
 import { JSX } from "react/jsx-runtime";
 import {
-  getTypeString,
-  getType,
   unpackArrayType,
   getChildPropertyDefinition,
 } from "../utils/astTypeHelpers";
 import { CodeTooltip } from "./CodeTooltip";
+import TreeNodeContainer, { TreeNodeContainerProps } from "./TreeNodeContainer";
 
 import { TypeMetadata } from "../utils/astTypeDefinitions";
 import { useCodeTranslationContext } from "../context/codeTranslationContext";
@@ -20,6 +18,10 @@ interface TreeNodeProps {
   expanded: boolean;
   onToggle: () => void;
   path: string;
+  renderChild: (
+    childProps: TreeNodeContainerProps,
+    key: string | number
+  ) => React.ReactNode;
   typeMetadata: TypeMetadata;
   searchTerm?: string;
   isDiffMode?: boolean;
@@ -43,6 +45,7 @@ export const TreeNode: React.FC<TreeNodeProps> = ({
   expanded,
   path,
   onToggle,
+  renderChild,
   typeMetadata,
   searchTerm = "",
   isDiffMode = false,
@@ -288,7 +291,11 @@ export const TreeNode: React.FC<TreeNodeProps> = ({
   // Render primitive values
   if (value === null || value === undefined) {
     return (
-      <div className={diffClassName} onMouseEnter={handleMouseEnter} data-testid={"node-" + path}>
+      <div
+        className={diffClassName}
+        onMouseEnter={handleMouseEnter}
+        data-testid={"node-" + path}
+      >
         {getRenderedContent(false, renderValueWithDiff("null"), false)}
       </div>
     );
@@ -299,7 +306,11 @@ export const TreeNode: React.FC<TreeNodeProps> = ({
     const displayValue =
       typeof value === "string" ? `"${value}"` : String(value);
     return (
-      <div className={diffClassName} onMouseEnter={handleMouseEnter} data-testid={"node-" + path}>
+      <div
+        className={diffClassName}
+        onMouseEnter={handleMouseEnter}
+        data-testid={"node-" + path}
+      >
         {/* include empty span to ensure indentation aligns with expandable nodes */}
         <span className="tree-arrow"></span>
         {getRenderedContent(false, renderValueWithDiff(displayValue), false)}
@@ -349,24 +360,22 @@ export const TreeNode: React.FC<TreeNodeProps> = ({
             // Pass through diff props for child nodes if they have diff annotations
             const childDiffProps = getChildDiffProps(value, index, item);
 
-            return (
-              <TreeNodeContainer
-                key={index}
-                nodeKey={`[${index}]`}
-                value={item}
-                level={level + 1}
-                parentInferredType={
-                  punctuatedType
-                    ? unpackArrayType(punctuatedType.type as string)
-                    : typeMetadata.type
-                    ? unpackArrayType(typeMetadata.type)
-                    : undefined
-                }
-                searchTerm={searchTerm}
-                path={`${path}.${index}`}
-                hiddenNodes={hiddenNodes}
-                {...childDiffProps}
-              />
+            return renderChild(
+              {
+                nodeKey: `[${index}]`,
+                value: item,
+                level: level + 1,
+                parentInferredType: punctuatedType
+                  ? unpackArrayType(punctuatedType.type as string)
+                  : typeMetadata.type
+                  ? unpackArrayType(typeMetadata.type)
+                  : undefined,
+                searchTerm: searchTerm,
+                path: `${path}.${index}`,
+                hiddenNodes: hiddenNodes,
+                ...childDiffProps,
+              },
+              index
             );
           })}
       </div>
@@ -431,114 +440,36 @@ export const TreeNode: React.FC<TreeNodeProps> = ({
             ? childPropertyDefinition.generic
             : childPropertyDefinition?.type;
 
-          return (
-            <TreeNodeContainer
-              key={key}
-              nodeKey={key}
-              value={childValue}
-              level={level + 1}
-              parentInferredType={parentInferredType}
-              searchTerm={searchTerm}
-              hiddenNodes={hiddenNodes}
-              path={`${path}.${key}`}
-              {...childDiffProps}
-            />
+          return renderChild(
+            {
+              nodeKey: `${key}`,
+              value: childValue,
+              level: level + 1,
+              parentInferredType: parentInferredType,
+              searchTerm: searchTerm,
+              path: `${path}.${key}`,
+              hiddenNodes: hiddenNodes,
+              ...childDiffProps,
+            },
+            key
           );
+
+          // return (
+          //   <TreeNodeContainer
+          //     key={key}
+          //     nodeKey={key}
+          //     value={childValue}
+          //     level={level + 1}
+          //     parentInferredType={parentInferredType}
+          //     searchTerm={searchTerm}
+          //     hiddenNodes={hiddenNodes}
+          //     path={`${path}.${key}`}
+          //     {...childDiffProps}
+          //   />
+          // );
         })}
     </div>
   );
 };
 
-// Container component that manages expand/collapse state for each node
-interface TreeNodeContainerProps {
-  nodeKey: string;
-  value: any;
-  level: number;
-  path: string;
-  searchTerm?: string;
-  parentInferredType?: string | string[];
-  isDiffMode?: boolean;
-  diffStatus?:
-    | "added"
-    | "nested-add"
-    | "removed"
-    | "updated"
-    | "unchanged"
-    | "contains-changes"
-    | "contains-nested-changes";
-  beforeValue?: any;
-  afterValue?: any;
-  hiddenNodes?: string[];
-}
-
-const TreeNodeContainer: React.FC<TreeNodeContainerProps> = (props) => {
-  // get all this metadata at tree node level; pass to TypeTooltip
-  const [type, kind] = React.useMemo(() => {
-    return getTypeString(props.value, props.nodeKey, props.parentInferredType);
-  }, [props.value, props.nodeKey, props.parentInferredType]);
-
-  const [typeDefinition, arrayType] = React.useMemo(() => {
-    if (type) {
-      return getType(type);
-    }
-    return [undefined, false];
-  }, [type]);
-
-  // do the above once each for old and new values on contains-changes nodes
-  const [prevType, prevKind] = React.useMemo(() => {
-    const childChanges = props.value ? props.value.childChanges : undefined;
-    if (childChanges && (childChanges._astType || childChanges.kind)) {
-      const hackVal = {
-        _astType: childChanges._astType ? childChanges._astType.oldValue : type, // might need to handle removed ast types here better; (generally scenarios other than UPDATE _astType)
-        kind: childChanges.kind ? childChanges.kind.oldValue : kind,
-      };
-      return getTypeString(hackVal, props.nodeKey, undefined);
-    }
-    return [type, kind]; // if _astType not in changes, return already computed (prevType is same as currentType)
-  }, [type, kind, props.value, props.nodeKey]);
-  const [prevTypeDefinition, prevArrayType] = React.useMemo(() => {
-    if (prevType) {
-      return getType(prevType);
-    }
-    return [undefined, false];
-  }, [prevType]);
-
-  const typeMetadata = {
-    type: type,
-    typeDefinition: typeDefinition,
-    arrayType: arrayType,
-    kind: kind,
-    prevType: prevType,
-    prevTypeDefinition: prevTypeDefinition,
-    prevArrayType: prevArrayType,
-    prevKind: prevKind,
-  };
-
-  const autoCollapse = props.isDiffMode
-    ? props.diffStatus === "unchanged" ||
-      props.diffStatus === "removed" ||
-      shouldAutoCollapse(type, typeDefinition)
-    : shouldAutoCollapse(type, typeDefinition);
-
-  const [expanded, setExpanded] = React.useState(!autoCollapse);
-
-  const handleToggle = () => {
-    setExpanded(!expanded);
-  };
-
-  // Hide nodes that are in the hidden nodes list (when filtering is active)
-  if (props.hiddenNodes && props.hiddenNodes.includes(props.nodeKey)) {
-    return null;
-  }
-
-  return (
-    <TreeNode
-      {...props}
-      expanded={expanded}
-      onToggle={handleToggle}
-      typeMetadata={typeMetadata}
-    />
-  );
-};
-
-export default TreeNodeContainer;
+export default TreeNode;
